Expand should be working but isn't
I added the expanded property to every object as it is created in the store.
Created an expanded button which toggles the value of the item being clicked.
I modified the generate bookmarks list items function to add or remove elements based on the item property.
The only thing is....it is not rendering to the page. I added a console log to confirm the correct string is being generated for the bookmarks and it is correct.